hal: fix usecase selection in adev_open_output_stream
During an active multichannel WFD session, YouTube
streaming stutters and audio is not routed to headset
when it is plugged in at the source or to the speaker
when the session is paused at the sink.

Multichannel playback usecase gets selected for YouTube
streaming when it is started during a WFD session. When
a headset is plugged in or the session is paused from the sink,
as there are no default mixer controls for multi-channel-playback,
there is an error when they are set.

Fix the issue by not selecting multi-channel-playback usecase
if DIRECT_PCM flag is present when the output is opened. This
ensures that compress-offload-playback2 usecase gets selected
and as there are valid mixer controls for speaker and headset for
this usecase, audio routes to the appropriate device as expected
when there is a device switch.

CRs-Fixed: 961105
Change-Id: I0bb638a4ac3f219d0a161dbfccb3aa540910dfc2