BACKPORT: hal: Fix failure in VtsHalAudioV2_0Target test due to spkr protection
1. Speaker protection cal file is not generated as
the path /data/misc/audio is not accessible. Change the
cal file path to /data/vendor/misc/audio to fix this.
2. VtsHalAudioV2_0Target fails to execute completely and results in
failure of target boot up after VTS execution as system server waits
indefinitely for audioserver to come up when speaker protection
is enabled.

This is due to improper clean up of speaker protection threads
and fd leaks.

Add spkr_prot_deinit function which joins the calibration thread
and cleans up the memory allocated for thread. Close the leaked
fds in the speaker calibration thread to ensure successful
execution of VTS.

CRs-Fixed: 2076733
Change-Id: I55a03dcafa3ed7bdf6f4d6d1bf72d84fbcc9c404
Signed-off-by: Julian Veit <claymore1298@gmail.com>